+++ moco-core/src/test/java/com/github/dreamhead/moco/MocoPortTest.java
@@ -4,9 +4,9 @@ import com.github.dreamhead.moco.helper.MocoTestHelper;
 import org.junit.Before;
 import org.junit.Test;
 
-import static com.github.dreamhead.moco.Moco.httpserver;
-import static com.github.dreamhead.moco.helper.RemoteTestUtils.root;
+import static com.github.dreamhead.moco.Moco.httpServer;
 import static com.github.dreamhead.moco.Runner.running;
+import static com.github.dreamhead.moco.helper.RemoteTestUtils.root;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -21,7 +21,7 @@ public class MocoPortTest {
 
     @Test
     public void should_create_server_without_specific_port() throws Exception {
-        final HttpServer server = httpserver();
+        final HttpServer server = httpServer();
         server.response("foo");
 
         running(server, new Runnable() {
@@ -34,7 +34,7 @@ public class MocoPortTest {
 
     @Test(expected = IllegalStateException.class)
     public void should_not_get_port_without_binding() throws Exception {
-        final HttpServer server = httpserver();
+        final HttpServer server = httpServer();
         server.port();
     }
 }